+++ SVRRoot/SVRBusiness/include/SVRParametersService.hpp
@@ -33,31 +33,31 @@ public:
     int remove_by_dataset(const bigint dataset_id);
 
     std::deque<datamodel::SVRParameters_ptr> get_all_by_dataset_id(const bigint dataset_id);
-    std::deque<datamodel::SVRParameters_ptr> get_by_dataset_column_level(const bigint dataset_id, const std::string &input_queue_column_name, const unsigned decon_level, const unsigned step);
+    std::deque<datamodel::SVRParameters_ptr> get_by_dataset_column_level(const bigint dataset_id, const std::string &input_queue_column_name, const uint16_t decon_level, const uint16_t step);
 
     static datamodel::t_param_set slice(
             const datamodel::t_param_set &params,
-            const unsigned chunk_ix = std::numeric_limits<unsigned>::max(),
-            const unsigned grad_ix = std::numeric_limits<unsigned>::max());
+            const uint16_t chunk_ix = std::numeric_limits<uint16_t>::max(),
+            const uint16_t grad_ix = std::numeric_limits<uint16_t>::max());
     static datamodel::t_param_set slice(
             const std::deque<datamodel::SVRParameters_ptr> &params,
-            const unsigned chunk_ix = std::numeric_limits<unsigned>::max(),
-            const unsigned grad_ix = std::numeric_limits<unsigned>::max());
+            const uint16_t chunk_ix = std::numeric_limits<uint16_t>::max(),
+            const uint16_t grad_ix = std::numeric_limits<uint16_t>::max());
 
-    static datamodel::t_param_set::iterator find(datamodel::t_param_set &params, const unsigned chunk_ix, const unsigned grad_ix);
-    static datamodel::t_param_set::const_iterator find(const datamodel::t_param_set &params, const unsigned chunk_ix, const unsigned grad_ix);
+    static datamodel::t_param_set::iterator find(datamodel::t_param_set &params, const uint16_t chunk_ix, const uint16_t grad_ix);
+    static datamodel::t_param_set::const_iterator find(const datamodel::t_param_set &params, const uint16_t chunk_ix, const uint16_t grad_ix);
     static datamodel::SVRParameters_ptr find_ptr(
             const datamodel::t_param_set &params,
-            const unsigned chunk_ix = std::numeric_limits<unsigned>::max(),
-            const unsigned grad_ix = std::numeric_limits<unsigned>::max());
+            const uint16_t chunk_ix = std::numeric_limits<uint16_t>::max(),
+            const uint16_t grad_ix = std::numeric_limits<uint16_t>::max());
 
-    static bool check(const datamodel::t_param_set &params, const unsigned num_chunks);
+    static bool check(const datamodel::t_param_set &params, const uint16_t num_chunks);
 
     static datamodel::SVRParameters_ptr is_manifold(const datamodel::t_param_set &param_set);
 
-    static unsigned get_trans_levix(const unsigned levels);
+    static uint16_t get_trans_levix(const uint16_t levels);
 
-    static std::set<unsigned> get_adjacent_indexes(const unsigned level, const double ratio, const int level_count);
+    static std::set<uint16_t> get_adjacent_indexes(const uint16_t level, const double ratio, const uint16_t level_count);
 };
 
 }
